Add files via upload
Created a new file that holds all the parameter values used in the study, and generates all the combinations that we ran trials for.  REAL_Simulation.cpp calls the lambda Get_Parameter_Combinations() to generate a global list of parameter combinations, which are then run by the simulation. I just wanted to move all parameters to a dedicated file so that the main sim file looks cleaner.